+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this package will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [1.0.0-pre.11] - 2025-05-29
+
+### Changed
+
+- Run Unity Hub if it isn't running when refreshing the cloud access token.
+- Show invalid cloud project banner if cloud settings not valid for 10 seconds or more.
+
 ## [1.0.0-pre.10] - 2025-05-27
 
 ### Changed